feat: implement internal exports handling
